Switch to polling in input thread.
This sucks, I was hoping to keep the interface blocking, but that's just not
possible. Why? Well, for one thing the SIGWINCH handler is fucked up such that
keys get dropped or mangled if you resize more than once without any input in
between (this commit fixes that) and, also, being able to stop or pause the
input_thread is integral for text browser support (the next series of patches).

So, tough luck =(.

Signed-off-by: Jack Miller <jack@codezen.org>