Handle M- C- keybinds.
So I spent *way* too long trying to get good ol' C-M- keys to work as well when
I realized that not only is it impossible to implement in a cross-terminal way (
C-M-x gets represented the same as M-x and M-C-x as C-x) but it doesn't even
work in old Canto. Derr.

tl;dr I don't care enough about hand-breaking C-M-x keybinds enough to look at
how Emacs handles it.

Signed-off-by: Jack Miller <jack@codezen.org>