Add line to reader output to compensate for trimming.
The reader code, like the rest of the gui elements, knows to disregard the
last line of a curses pad because it's bugged (rather it has a legacy
inability to write to the bottom right corner). This patch reconciles the
reader renderer with that assumption, which the rest of the code was already
making.

Signed-off-by: Jack Miller <jack@codezen.org>