Use new per-socket transforms. Better refresh.
This uses the new per-socket transforms to allow the temp-transform command.
This will eventually be obsoleted for things like search and the transform
cycle.

In addition, because sorts are not reflected just by the addition /
subtraction of items, we reset all tags on refresh. This not only gives much
better (read: any at all) visual feedback on refresh, but also allows sorts
to be applied immediately.

Signed-off-by: Jack Miller <jack@codezen.org>