Add symbols to tag/item range commands.
This adds two special range symbols to any command that can take a list of
items or tags. The first symbol is $ which represents the last item/tag in
the range. The second symbol is . which represents the current item/tag if
there's a selection (defaults to 0 if no selection is made). 0 will always be
the first item in any range so no special symbol is necessary.

This gives you the ability to do nice things like

:item-state read 0-.

To mark all items up to and including the selection read while leaving all
items afterwards untouched.

Commit soon will allow you to use the range on items in the tag range.

Signed-off-by: Jack Miller <jack@codezen.org>