Update all item list commands to be possibly relative.
This is a major update to every command that takes a list of items. You can
now optionally restrict your list to a single tag.

For example:

:item-state read 1 3 4 7

As usual this will mark items 1, 3, 4, and 7 read

:item-state read t:0 1 3

This will mark the first *visible* feed's second and fourth items read.

:item-state read T:0 1 3

This will mark the first *absolute* feed's second and fourth item read.
Absolute in this context means the first feed defined in your config,
regardless of whether it's visible or not.

In addition, these all support the usual complete / partial prompting, so:

:goto t:

Will prompt you for a single tag and then items in that tag, with the
appropriate item enumeration.

Signed-off-by: Jack Miller <jack@codezen.org>