Hook based item add / remove
This allows us to get the offset / selected item messing about out of
gui.py, where it didn't belong. It also allows us the ability to
remember the selected item and re-select it if it reappears (and there's
no selection already made).

Not sure if it's good behavior yet though, if you're on a read item and
refresh (with the filter_read transform), you'll get put back to the
unselected state...

Signed-off-by: Jack Miller <jack@codezen.org>