Add story format string
The format of the story representation is now based on a format string
instead of hard-coded in python.

The format string supports a basic grammar similar to that found in NRSS
and vim.

There are three levels of evaluation.

-----------------
The first parses conditionals, which are used like ternaries. For
example:

%?{en}([%i] :)

Will evaluate whether the conditional "en" is true, and if it is, render
"[%i] ". Otherwise it will render nothing. Ternaries can, in turn, be
nested, although the default format string doesn't use this:

%?{en}(%{sel}(en and sel:en not sel):not en)

-----------------
The second level of evaluation takes any further %{} or single escapes
(like %t for title), evaluates them and adds them to the representation.

-----------------
The final level of evaluation is done on render, which takes the various
style / color escapes ( %B %b %U %u %S %s %D %d %R %r %0-8 ). These
escapes are ignored through the previous steps.

NOTE: Any format brackets ( {} ) will be evaluated in a limited Python
environment. So, for example, given that "item" refers to the current
item being rendered, you can do things like %?{item.selected} for a
ternary or %{item.annotate()} if you've added a function to stories with
a plugin.