Massive drawing rework.
The previous commit added the ability to use a format string to define the
story's representation. This is definitely a priority, the sole feature from
NRSS that was dropped in old Canto that shouldn't have been (I foolishly
thought that being able to define new looks in Python would be a standard use
case, but format strings are 10x easier).

The previous drawing system was O(n) where n is the number of items being
tracked by the interface. This is fine when you have <=100 items and 100%
built-in representation, even with 500 items a decent processor would only
slow a little (although even that's ridiculous for a curses app).

The new drawing system is O(n) where n is the number of lines on your screen.
On my 1920x1080 screens, using xterm that means about 65 items for a
screenful and that never changes, regardless of how far up or down in the
list you are or how many items the interface is tracking.

This along should make performance skyrocket. In addition, however, I used
this rework as an opportunity to add a bit of reactive programming to the
drawing system and make it very lazy. That means that instead of re-rendering
all of the items on redraw, the redraw is now just a copy operation. This is
accomplished because the items will update their representation immediately
when it's possible that it's changed. At this stage, the only unfortunate
part about this is that just starting the interface with 1k items takes
noticeably longer than on the old codebase. Same with operations that effect
all rendered stories or the order of stories (like enumeration or promote).

With the new structure of the on-screen elements (tags, stories, and the
reader), it would be trivial, if necessary, to implement even further
laziness. For example, the elements marking themselves as needing update and
only executing that update if they are on-screen. I haven't quite decided if
it's worth the complexity but I designed with that possibility in mind.

Signed-off-by: Jack Miller <jack@codezen.org>