Even more drawing efficiency.
As I mentioned in the previous major drawing commit, I designed it with
the idea that it could be made even lazier if necessary. Well, testing
the intervening border patch I realized that causing a resize of the
base taglist window would cause a full item refresh (which takes a
second or two), which kinda sucks when, for example, the reader isn't
floating (i.e. whenever you open the reader, it causes a taglist
resize).

So, this patch goes the extra mile and changes taglist render such that
it will only cause a refresh for visible items. This was accomplished by
rephrasing all of the scrolling options to describe a target object
(which is the selection or the first visible item), and a target window
offset. The redraw() code will then extrapolate the remaining window
contents from that object and offset, updating as needed.

This means any operation, regardless of how many items it truly effects
will, at most, cause a screenful of items to actually update their
curses pads.