Move tick() out of signal thread context.
Similar to disconnect, reconnect, and winch we should put an item on the
queue and then let it be handled by the main gui thread. This probably
didn't have any bugs considering it's the only code that was touching
gui.updates and it was just writing to the socket, but it was bad
practice.

Signed-off-by: Jack Miller <jack@codezen.org>