Better signal handlers.
Return to the philosophy of merely setting flags. Even logging and Queue
manipulation isn't guaranteed to be happy from a signal context. While
we do it for alarm (tick) and winch, go ahead and flag-up
disconnect/reconnect just because they're coming from the connection
thread rather than the gui thread.

Signed-off-by: Jack Miller <jack@codezen.org>