C-C test harness initial commit.
This is similar to the black box testing of c-d. I don't really care
about the internal state as much as I want to be able to verify that
everything looks right when displayed and doesn't except.

As of now, this is a bit kludgy as it expects to be able to set the size
of the xterm directly. This isn't usually a problem but in Xmonad (my
WM), the geometry is ignored. I worked around this by adding a window
filter to ignore windows with the title: canto-test, but it will break
in a tiling WM just because the geometry is wrong.

This could be worked around by setting the geometry internally, but then
you still have to deal with the case where the window is even smaller
than that and it would change the code a bit which this test harness is
explicitly trying to avoid.

Also note that this dumps the windows of all top-level gui window
objects, not necessarily just what is displayed on screen.

Final note is that, like the c-d harness, this will only test the
version of Canto installed and accessible in your path not necessarily
the version in the git repo.

Signed-off-by: Jack Miller <jack@codezen.org>