Re-introduce collapse (default to toggle: c)
This shouldn't have been as invasive as it was, but the selection code
had to be changed to deal with the case in which a selection was a tag,
as well as some tag opt errors that occurred when adding the new option
to previously defined tag sections.

I'm fairly certain I've dealt with all of the collapse cases, but I
could be wrong.

Signed-off-by: Jack Miller <jack@codezen.org>