Implement update.style
update.style can be set to "maintain" or "append"

Append mode is the default, and has been since the beginning. In this
mode, new items are appended to the end of the tag. This is nice because
it allows the user to know that all items above the cursor have been
read without necessarily having to set them as read as he progresses.
The downside is that updates can leave a tag out of expected order.
(ABCD -> update -> ABCDAB instead of AABBCD).

Maintain mode is a new addition that will reorder the items based on the
order that they were given by the daemon. This means new items may
appear above the cursor, but it also means that sort order is maintained
(thus the name).

Signed-off-by: Jack Miller <jack@codezen.org>