Lazily prep strings for display.
Instead of erroneously grabbing all string attributes and preparing them
to be displayed, only escape them etc. immediately before they're
actually going to be put on the string.

This is the alpha variant of the fix for old canto issue #21 (%s in
links getting screwed up on the way to the browser).

As a side effect, any string OTHER THAN THE STRINGS PROVIDED, evaluated
in any format string needs to have prep() called on it, which is now
exported to the evaluation function.

So, for example, having

"%{item['content']['nondefault']}"

You'd now have to call prep on it:

"%{prep(item['content']['nondefault'])}"

The old way will work still, but any character entities, slashes, or %s
will be interpreted.

This is actually more consistent in general, considering that the
content being auto-escaped was only values that are superficially
strings (i.e. it didn't prepare strings in data structures or anything).

Signed-off-by: Jack Miller <jack@codezen.org>