Add support for in-interface color changes on the fly
The basic format of the command is

:color [pair] [fg] (bg)

Where pair is either a number 0 < x <= curses COLOR_PAIRS (1-8 typically),
or "deffg" or "defbg". There are some aliases for the default theme as
well ("main","unread","read","link","image","quote"). fg and bg are
numeric curses colors 0 <= x < COLORS or an English color as defined by
curses ("black","blue","cyan","green","magenta","red","white","yellow")
in addition to "pink" (an alias for magenta) and also -1 which means it
should default to deffg/defbg.

So, typically, you'll do

:color main blue

To change main (text and tag header) to blue text on the same
background.

:color main -1 white

To change main to same foreground on a white background etc.

Signed-off-by: Jack Miller <jack@codezen.org>