Be more useful when disconnected.
This keeps the disconnected message from re-displaying every time you
perform an action while disconnected. It also short-circuits such that
the writes never occur (but are mentioned in the log as discarded).

This is useful when the daemon is no longer available and cannot be
:reocnnect'd. You can at least open what links you have in the browser,
although the reader won't be happy not being able to fetch the rest of
the info.

TBD: If it's worth it to add a :sync (or other command) that would copy
the state of the items in the client to a server. The thought being that
you could be disconnected, read a bunch of items, reconnect, and then
tell the daemon about the read items.

Signed-off-by: Jack Miller <jack@codezen.org>