Add "fetch" capability to reader
Fetch differs from goto only in that it creates a temporary directory
(/tmp/canto-*), downloads the target file with the same filename into
that directory and then invokes the given browser on it.

This is intended to ease the use of more advanced scripts on the target
file. For example, having your browser set to a MIME aware opening
program (mimeo, mimeopen, etc.) or even just an extension aware
(xdg-open) program. The URL as included in the RSS will have a
persistent mimetype of HTTP(S) which is correct, but clearly not
interesting.

The files and temporary directories are cleaned up on exit with a new
exit hook.

The default reader key for fetch is, you guessed it, 'f'.

Signed-off-by: Jack Miller <jack@codezen.org>