Correct reader_item unset
reader_item must be unset to trigger the loss of protection in the
daemon. Previously we were doing this on die() which is called under all
circumstances that the window would need to be recreated, including
resize, so unsetting there causes c-c to lose the reader_item on resize,
which is stupid.

Now we override the GuiBase cmd_destroy() which is only called when the
user has dismissed the window.

Signed-off-by: Jack Miller <jack@codezen.org>