Avoid replacing unknown charrefs
Initially I considered that converting unknown charrefs into [?] would
basically let the user know that some character is missing. Now I
believe that the most likely cause of a failed ref is that an & is
unhandled (i.e. in plaintext feeds) so this ends up eating the next
characters after it.

This patch should fix the plaintext case (common) and punish the unknown
charref case by letting the bad ref leak into the interface. I think
this is preferable to breaking the plaintext case to better handle the
rare bad ref.

Signed-off-by: Jack Miller <jack@codezen.org>