Add cursor behavior options
Similar to old canto there are three options with multiple values.

:cursor_type = [edge|top|middle|bottom]

Determines where the cursor is pinned (or not pinned at all, with edge)

:cursor_scroll = [scroll|page]

Determines what happens when the interface scrolls. Scroll will show one
new item, page will show a page full of new items.

:cursor_edge = [integer]

Determines the margin from the top or bottom for scrolling. The default,
5, from old Canto, will cause the interface to scroll if the new item is
within 5 lines of the top or bottom. When used with type = top or
bottom, it determines the absolute distance from the top or bottom of
the screen.

All of these can be set independently.

Signed-off-by: Jack Miller <jack@codezen.org>