Fix hung items
Not really hung, per se, although they appeared that way. Canto protects
items that you've selected or are reading so that they don't disappear
right from under you and cause your cursor to move without you. As a
side-effect though, an item that *should* be filtered out might not be,
but then unless that tag gets updated by the daemon, or you change and
attribute in that tag, nothing is going to cause that tag to ever get
updated. So, especially on slow moving feeds, the item will just stay
there forever thanks to our lazy update cycle.

We fix this by causing the client to view changing protection (i.e.
select / unselect / reader / unreader) as grounds to ask for items on
tick.

Signed-off-by: Jack Miller <jack@codezen.org>