Create priority thread.
With the daemon round-robining connections, we can start a second
connection that will only be used for stuff that the user is waiting on.
For now, this is just reader content, which used to get stuck for
seconds at a time waiting for background updates to process.

In the future, it could perhaps be expanded to filter settings, although
it wouldn't eliminate getting ITEMS responses processed without them
set.

Signed-off-by: Jack Miller <jack@codezen.org>