Tie sync into gui thread
I didn't want to have to do this because the sync is a little heavier
weight, but it's the easiest solution to the problem that some commands
(i.e. ones that use items or function on ranges of items) require the
GUI stuff (like TagList's linked list of objects) to be up to date after
a sync.

If the work ends up being too much, I'll do all the prep before and turn
sync() into a simple reference switch. For now, I don't notice a
difference.