Working completion snapshot
Relative scrolling works as a proof of concept.

The root commands and args should be completed and validated, but
considering rel-set-cursor takes a single int argument and can't really
say it's working.

The next step is to convert the existing commands to register and use
registered argument types.

Doing something intelligent with out-of-line completion help text is
also a priority. I might change the help text to a generator as well.