Remote command hookup
:remote [remote-cmd] [args]

Basic pass through for canto-remote. It will complete the command, but
not the args for it because usually there are no args (e.g. listfeeds),
or the args are totally incompletable (e.g. addfeed). In addition, most
of the remote functionality a user should get should be wrapped, (i.e.
'add' and 'del' instead of 'remote addfeed ...' or 'remote delfeed ...'