Fix triple-tab readline screen corruption.
Okay, this *really* sucks, but it's become very apparent that the Python
readline module was never really supposed to be anything but a binding
for non-curses consoles.

In this case, readline has a bunch of cases where it prints to the
screen itself which is *awful* behavior for a library.

Fortunately, readline is good enough to allow callers to override
rl_outstream to discard this (and any other) output.

Unfortunately, the readline-enhanced Python input function manually
overrides that setting with stdin/stdout.

So, for now my solution is to basically bastardize the built-in
PyOS_Readline function to do pretty much the same thing, except
redirecting rl_outstream to /dev/null.

CAVEATS:

- This is hugely ugly, but I'm having trouble seeing how I can avoid it
  since I have no external control over what the Python module does on
input()

- It's causing the client to hang on a signal on exit, but this is a
  problem that existed before and is now consistent. Giving another
signal, like WINCH or INT jars it lose. This will probably be my next
problem.