Add tag-item command
Add or remove an arbitrary user:tag.

I've added a small change to the story format so if you

:tag-item favorite .

for example, you can see a # appear in front of the story and

:tag-item -favorite .

will cause it to disappear

It's a bit rough and I'm sure there are bugs. For example, we probably
shouldn't get a message box for adding user tags...