Propagate die() on deleted tag
During the normal course of operation, story.die() is called as part of
sync(), but when a tag dies (on DELTAG), this is ignored. This causes
the hooks to be uncleaned up and the hook system to hold references to
story objects. This fixes a small memory leak caused by deleting tags.