List keybinds on the basic :help page
Command list moved to :help commands

This is still rough because it shows the raw commands instead of natural
English descriptions. Sometimes that works out (search), sometimes it
doesn't (foritems & goto & item-state read & clearitems).

Not quite sure how I'm going to deal with that yet, but it might be part
of the aliases in the command system (i.e. rename "goto" to something
else, and then alias "goto" to the new string of commands).