Config setting overhaul
A lot of small changes adding up to a big change.

The big change is that you can now do

    :set [option] [value] (set a value)
    :set [option]         (display current value)

For pretty much everything in the configuration.

This works for c-c settings, as well as daemon settings, so

    :set feed.(whatever)
    :set tag.(whatever)
    :set defaults.(whatever)

will now use the selection, if necessary, to Do The Right Thing

NOTE: This is *not* intended to be the way to change most settings. Most
things, like binds, and colors, and all the various boolean toggles are
properly attached to better helpers or keybind commands. This *is*
intended to provide a way to manipulate the configuration for simple
stuff like "browser".

Now for the list of small changes

- Raw config commands (like :global_transform, :browser, and
  :txt_browser) have now been removed in favor of "set
global_transform", "set browser", "set txt_browser" which are in turn
aliases for "set defaults.global_transform" "set browser.path" and "set
browser.text" for some level of backwards compatibility.

- CantoCurses.tag settings are now CantoCurses.tagobj settings (to make
  way for tag.whatever to work intuitively.

- c-c now tracks daemon defaults and feed settings, and provides an
  interface to them, but doesn't validate them at all (that's up to the
daemon).

- :tag-config has been eliminated in favor of "set tag.whatever"

- Config settings are no longer abusing :remote

- The config module now has an 'eval_settings' list of uncompiled
  regexes. Config options matching any of these regexes are eval()'d in
an empty environment before they're set. Plugins can manipulate this
when they're first evaluated, but immediately before the curses_start
hook, they're compiled and should not be touched any longer.