Minor fix to thread disconnect message
I thought about it a bit, and I think this is probably fine for
disconnect support. The interface remains as responsive as possible (you
can still scroll and go to links, etc.). Requests for new information
will obviously fail, but that's handled gracefully in testing.

Implementing :reconnect is a possibility, but not a priority. Especially
when restarting the client will either restart the daemon (local socket)
and do the right thing, or it won't (remote socket) at which point
:reconnect would also fail and the user has to go manually restart it
anyway.

tl;dr a warning that your state will be forgotten is as far as I'm going
to bother with disconnect for 0.9.0