Add "smartlink" plugin for advanced link fetching
- Can specify handlers for specific content either based on URL, or
  based on the output of `file` when a temporary file is downloaded.

- Regex or simple matches.

- Text and graphical handlers ok.

CAVEATS: Right now, text/"pause" handlers are a bit broken because the
pause and unpause callbacks don't work outside of the user input thread
(fetching occurs in its own thread so the interface is responsive while
a potentially slow download occurs). I'm going to have to come up with a
way to wake up the user input thread to get it to stop consuming keys,
or to pass stdin directly to the child process.