Add cleantitle plugin + Story support
This plugin adds the ability to do simple text replacement on story
titles before they're displayed. Useful for stripping out content that
(according to the feed) should be there, but (according to the user) is
crap. Default case, newlines and some HTML tags that seem to get left
behind occasionally.