config prot_thread can neither set or clear processed flag
This wait would obviously deadlock since the prot_thread is what has to
actual process the PONG and set flag. The good news, however, is that
the config prot_thread is only used on NEW and DELTAGS causing config
changes and any writes stemming from those are pretty much just
informative to the daemon, whereas this mechanism is intended for user
changes being reflected before subsequent requests are made.