Fix utterly broken tag core new/cur/old detection
Turns out when you double sort one list instead of sorting both lists,
assuming that both lists are sorted is a bad idea.

Fortunately, we only used this logic to notify other objects, so the
worst effect is that we woke up the gui thread for nothing, and the
pending item count got out of wack, but in theory it could also remove
items improperly, although I haven't seen that in the wild.

Closes #40