Don't re-release gui thread on SIGWINCH
While Python Events are obviously thread safe, they're not reentrant so
from a signal context we have to make sure that if we receive a new
signal before we finish processing the old one, we don't attempt to
set() again and deadlock.

This fixes canto-curses locking when you resize your terminal a few
hundred times a second.