Fix input_box refresh() out of order resize
Because readline is in charge of calling redisplay, and appears to call
redisplay on resize, and redisplay call input_box.refresh, which is
racing with potentially initing a new input_box, so if the input box's
init() hasn't been called yet, don't do anything.

Fixes exceptions when resizing frequently with a prompt open.